pyproject.toml
[project]
name = "academic-research-paper-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "pymupdf>=1.27.2",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
]
